requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.2.0
braceexpand==0.1.7
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
import_expression==1.1.5
jishaku==2.5.2
multidict==6.0.5
pyotp==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
resend==2.2.0
sib-api-v3-sdk==7.6.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.43.0
yarl==1.9.4
